Use server per test instead of OnceCell
