+++ utils/ogmios-client/tests/query_ledger_state.rs
@@ -1,50 +1,18 @@
 #![cfg(feature = "jsonrpsee-client")]
 
-use jsonrpsee::{http_client::HttpClient, server::Server, RpcModule};
+use jsonrpsee::http_client::HttpClient;
 use ogmios_client::{
 	query_ledger_state::{EpochBoundary, EpochParameters, EraSummary, QueryLedgerState},
 	types::{SlotLength, TimeSeconds},
 };
 use serde_json::json;
-use std::net::SocketAddr;
-use tokio::sync::OnceCell;
 
-const SERVER_ADDRESS: OnceCell<SocketAddr> = OnceCell::const_new();
-
-async fn get_server_address() -> SocketAddr {
-	SERVER_ADDRESS
-		.get_or_init(|| async { run_server().await.unwrap() })
-		.await
-		.clone()
-}
+mod server;
 
 #[tokio::test]
 async fn era_summaries() {
-	let address = get_server_address().await;
-	let client = HttpClient::builder().build(format!("http://{address}")).unwrap();
-	let era_summaries = client.era_summaries().await.unwrap();
-	assert_eq!(era_summaries.len(), 3);
-	assert_eq!(
-		era_summaries[0],
-		EraSummary {
-			start: EpochBoundary { time: TimeSeconds { seconds: 0 }, slot: 0, epoch: 0 },
-			end: EpochBoundary { time: TimeSeconds { seconds: 1728000 }, slot: 86400, epoch: 4 },
-			parameters: EpochParameters {
-				epoch_length: 21600,
-				slot_length: SlotLength { milliseconds: 20000 },
-				safe_zone: 4320
-			}
-		}
-	)
-}
-
-async fn run_server() -> anyhow::Result<SocketAddr> {
-	let server = Server::builder().build("127.0.0.1:0".parse::<SocketAddr>()?).await?;
-	let mut module = RpcModule::new(());
-	module.register_method("queryLedgerState/eraSummaries", |_, _, _| {
-		// This is real answer from Ogmios, trimmed down to 3 eras for brevity
-		json!([
-		  {
+	let address = server::for_single_test("queryLedgerState/eraSummaries", |_| {
+		Ok(json!([{
 			"start": {
 			  "time": {
 				"seconds": 0
@@ -113,11 +81,23 @@ async fn run_server() -> anyhow::Result<SocketAddr> {
 			  "safeZone": 129600
 			}
 		  }
-		])
-	})?;
-	let addr = server.local_addr()?;
-	let handle = server.start(module);
-	// It will stop when test main exists.
-	tokio::spawn(handle.stopped());
-	Ok(addr)
+		]))
+	})
+	.await
+	.unwrap();
+	let client = HttpClient::builder().build(format!("http://{address}")).unwrap();
+	let era_summaries = client.era_summaries().await.unwrap();
+	assert_eq!(era_summaries.len(), 3);
+	assert_eq!(
+		era_summaries[0],
+		EraSummary {
+			start: EpochBoundary { time: TimeSeconds { seconds: 0 }, slot: 0, epoch: 0 },
+			end: EpochBoundary { time: TimeSeconds { seconds: 1728000 }, slot: 86400, epoch: 4 },
+			parameters: EpochParameters {
+				epoch_length: 21600,
+				slot_length: SlotLength { milliseconds: 20000 },
+				safe_zone: 4320
+			}
+		}
+	)
 }
